feat: ci matrix tests
