+++ composer.json
@@ -15,6 +15,9 @@
             "Appwrite\\Spec\\": "src/Spec"
         }
     },
+    "autoload-dev": {
+        "psr-4": {"Tests\\": "tests"}
+    },
     "require": {
         "php": ">=7.0.0",
         "ext-curl": "*",
@@ -24,7 +27,7 @@
         "matthiasmullie/minify": "^1.3"
     },
     "require-dev": {
-        "phpunit/phpunit": "^7.0"
+        "phpunit/phpunit": "^9.5.13"
     },
     "minimum-stability": "dev",
     "prefer-stable": true
